improve performance
-reduce IFluidStateCapability ram usage by storing serialized long values instead of BlockPos ones
-ASMHooks#getFluidExtendedState now calls a lot less world based accessors

+++ src/main/java/git/jbredwards/fluidlogged_api/mod/common/message/FluidStateMessage.java
@@ -74,7 +74,7 @@ public final class FluidStateMessage implements IMessage
                     final FluidState fluidState = FluidState.deserialize(message.state);
 
                     //send changes to client
-                    cap.setFluidState(pos, fluidState);
+                    cap.setFluidState(pos.toLong(), fluidState);
 
                     //re-render block
                     FluidloggedUtils.relightFluidBlock(world, pos, fluidState);
